+++ src/oci/file_storage/models/create_export_details.py
@@ -65,8 +65,12 @@ class CreateExportDetails(object):
                  {
                     \"source\" : \"0.0.0.0/0\",
                     \"requirePrivilegedSourcePort\" : false,
-                    \"access\" : \"READ_WRITE\",
-                    \"identitySquash\" : \"NONE\"
+                    \"access\": \"READ_WRITE\",
+                    \"identitySquash\": \"NONE\",
+                    \"anonymousUid\": 65534,
+                    \"anonymousGid\": 65534,
+                    \"isAnonymousAccessAllowed\": false,
+                    \"allowedAuth\": [\"SYS\"]
                   }
                ]
 
@@ -98,8 +102,12 @@ class CreateExportDetails(object):
                  {
                     \"source\" : \"0.0.0.0/0\",
                     \"requirePrivilegedSourcePort\" : false,
-                    \"access\" : \"READ_WRITE\",
-                    \"identitySquash\" : \"NONE\"
+                    \"access\": \"READ_WRITE\",
+                    \"identitySquash\": \"NONE\",
+                    \"anonymousUid\": 65534,
+                    \"anonymousGid\": 65534,
+                    \"isAnonymousAccessAllowed\": false,
+                    \"allowedAuth\": [\"SYS\"]
                   }
                ]