+++ src/oci/management_dashboard/models/create_management_dashboard_details.py
@@ -94,6 +94,10 @@ class CreateManagementDashboardDetails(object):
             The value to assign to the parameters_config property of this CreateManagementDashboardDetails.
         :type parameters_config: list[object]
 
+        :param features_config:
+            The value to assign to the features_config property of this CreateManagementDashboardDetails.
+        :type features_config: object
+
         :param drilldown_config:
             The value to assign to the drilldown_config property of this CreateManagementDashboardDetails.
         :type drilldown_config: list[object]
@@ -127,6 +131,7 @@ class CreateManagementDashboardDetails(object):
             'type': 'str',
             'is_favorite': 'bool',
             'parameters_config': 'list[object]',
+            'features_config': 'object',
             'drilldown_config': 'list[object]',
             'freeform_tags': 'dict(str, str)',
             'defined_tags': 'dict(str, dict(str, object))'
@@ -152,6 +157,7 @@ class CreateManagementDashboardDetails(object):
             'type': 'type',
             'is_favorite': 'isFavorite',
             'parameters_config': 'parametersConfig',
+            'features_config': 'featuresConfig',
             'drilldown_config': 'drilldownConfig',
             'freeform_tags': 'freeformTags',
             'defined_tags': 'definedTags'
@@ -176,6 +182,7 @@ class CreateManagementDashboardDetails(object):
         self._type = None
         self._is_favorite = None
         self._parameters_config = None
+        self._features_config = None
         self._drilldown_config = None
         self._freeform_tags = None
         self._defined_tags = None
@@ -232,7 +239,7 @@ class CreateManagementDashboardDetails(object):
     def provider_name(self):
         """
         **[Required]** Gets the provider_name of this CreateManagementDashboardDetails.
-        Name of the service (for example, Logging Analytics) that owns the dashboard.
+        The user friendly name of the service (for example, Logging Analytics) that owns the dashboard.
 
 
         :return: The provider_name of this CreateManagementDashboardDetails.
@@ -244,7 +251,7 @@ class CreateManagementDashboardDetails(object):
     def provider_name(self, provider_name):
         """
         Sets the provider_name of this CreateManagementDashboardDetails.
-        Name of the service (for example, Logging Analytics) that owns the dashboard.
+        The user friendly name of the service (for example, Logging Analytics) that owns the dashboard.
 
 
         :param provider_name: The provider_name of this CreateManagementDashboardDetails.
@@ -256,7 +263,7 @@ class CreateManagementDashboardDetails(object):
     def provider_version(self):
         """
         **[Required]** Gets the provider_version of this CreateManagementDashboardDetails.
-        Version of the service that owns the dashboard.
+        The version of the metadata of the provider. This is useful for provider to version its features and metadata. Any newly created saved search (or dashboard) should use providerVersion 3.0.0.
 
 
         :return: The provider_version of this CreateManagementDashboardDetails.
@@ -268,7 +275,7 @@ class CreateManagementDashboardDetails(object):
     def provider_version(self, provider_version):
         """
         Sets the provider_version of this CreateManagementDashboardDetails.
-        Version of the service that owns the dashboard.
+        The version of the metadata of the provider. This is useful for provider to version its features and metadata. Any newly created saved search (or dashboard) should use providerVersion 3.0.0.
 
 
         :param provider_version: The provider_version of this CreateManagementDashboardDetails.
@@ -424,7 +431,7 @@ class CreateManagementDashboardDetails(object):
     def metadata_version(self):
         """
         **[Required]** Gets the metadata_version of this CreateManagementDashboardDetails.
-        Version of the metadata.
+        The version of the metadata defined in the API. This is maintained and enforced by dashboard server. Currently it is 2.0.
 
 
         :return: The metadata_version of this CreateManagementDashboardDetails.
@@ -436,7 +443,7 @@ class CreateManagementDashboardDetails(object):
     def metadata_version(self, metadata_version):
         """
         Sets the metadata_version of this CreateManagementDashboardDetails.
-        Version of the metadata.
+        The version of the metadata defined in the API. This is maintained and enforced by dashboard server. Currently it is 2.0.
 
 
         :param metadata_version: The metadata_version of this CreateManagementDashboardDetails.
@@ -635,6 +642,30 @@ class CreateManagementDashboardDetails(object):
         :type: list[object]
         """
         self._parameters_config = parameters_config
+
+    @property
+    def features_config(self):
+        """
+        Gets the features_config of this CreateManagementDashboardDetails.
+        Contains configuration for enabling features.
+
+
+        :return: The features_config of this CreateManagementDashboardDetails.
+        :rtype: object
+        """
+        return self._features_config
+
+    @features_config.setter
+    def features_config(self, features_config):
+        """
+        Sets the features_config of this CreateManagementDashboardDetails.
+        Contains configuration for enabling features.
+
+
+        :param features_config: The features_config of this CreateManagementDashboardDetails.
+        :type: object
+        """
+        self._features_config = features_config
 
     @property
     def drilldown_config(self):