+++ src/oci/management_dashboard/models/management_dashboard_for_import_export_details.py
@@ -98,6 +98,10 @@ class ManagementDashboardForImportExportDetails(object):
             The value to assign to the parameters_config property of this ManagementDashboardForImportExportDetails.
         :type parameters_config: list[object]
 
+        :param features_config:
+            The value to assign to the features_config property of this ManagementDashboardForImportExportDetails.
+        :type features_config: object
+
         :param drilldown_config:
             The value to assign to the drilldown_config property of this ManagementDashboardForImportExportDetails.
         :type drilldown_config: list[object]
@@ -132,6 +136,7 @@ class ManagementDashboardForImportExportDetails(object):
             'is_favorite': 'bool',
             'saved_searches': 'list[ManagementSavedSearchForImportDetails]',
             'parameters_config': 'list[object]',
+            'features_config': 'object',
             'drilldown_config': 'list[object]',
             'freeform_tags': 'dict(str, str)',
             'defined_tags': 'dict(str, dict(str, object))'
@@ -158,6 +163,7 @@ class ManagementDashboardForImportExportDetails(object):
             'is_favorite': 'isFavorite',
             'saved_searches': 'savedSearches',
             'parameters_config': 'parametersConfig',
+            'features_config': 'featuresConfig',
             'drilldown_config': 'drilldownConfig',
             'freeform_tags': 'freeformTags',
             'defined_tags': 'definedTags'
@@ -183,6 +189,7 @@ class ManagementDashboardForImportExportDetails(object):
         self._is_favorite = None
         self._saved_searches = None
         self._parameters_config = None
+        self._features_config = None
         self._drilldown_config = None
         self._freeform_tags = None
         self._defined_tags = None
@@ -239,7 +246,7 @@ class ManagementDashboardForImportExportDetails(object):
     def provider_name(self):
         """
         **[Required]** Gets the provider_name of this ManagementDashboardForImportExportDetails.
-        Name of the service (for example, Logging Analytics) that owns the dashboard.
+        The user friendly name of the service (for example, Logging Analytics) that owns the dashboard.
 
 
         :return: The provider_name of this ManagementDashboardForImportExportDetails.
@@ -251,7 +258,7 @@ class ManagementDashboardForImportExportDetails(object):
     def provider_name(self, provider_name):
         """
         Sets the provider_name of this ManagementDashboardForImportExportDetails.
-        Name of the service (for example, Logging Analytics) that owns the dashboard.
+        The user friendly name of the service (for example, Logging Analytics) that owns the dashboard.
 
 
         :param provider_name: The provider_name of this ManagementDashboardForImportExportDetails.
@@ -263,7 +270,7 @@ class ManagementDashboardForImportExportDetails(object):
     def provider_version(self):
         """
         **[Required]** Gets the provider_version of this ManagementDashboardForImportExportDetails.
-        Version of the service that owns the dashboard.
+        The version of the metadata of the provider. This is useful for provider to version its features and metadata. Any newly created saved search (or dashboard) should use providerVersion 3.0.0.
 
 
         :return: The provider_version of this ManagementDashboardForImportExportDetails.
@@ -275,7 +282,7 @@ class ManagementDashboardForImportExportDetails(object):
     def provider_version(self, provider_version):
         """
         Sets the provider_version of this ManagementDashboardForImportExportDetails.
-        Version of the service that owns the dashboard.
+        The version of the metadata of the provider. This is useful for provider to version its features and metadata. Any newly created saved search (or dashboard) should use providerVersion 3.0.0.
 
 
         :param provider_version: The provider_version of this ManagementDashboardForImportExportDetails.
@@ -431,7 +438,7 @@ class ManagementDashboardForImportExportDetails(object):
     def metadata_version(self):
         """
         **[Required]** Gets the metadata_version of this ManagementDashboardForImportExportDetails.
-        Version of the metadata.
+        The version of the metadata defined in the API. This is maintained and enforced by dashboard server. Currently it is 2.0.
 
 
         :return: The metadata_version of this ManagementDashboardForImportExportDetails.
@@ -443,7 +450,7 @@ class ManagementDashboardForImportExportDetails(object):
     def metadata_version(self, metadata_version):
         """
         Sets the metadata_version of this ManagementDashboardForImportExportDetails.
-        Version of the metadata.
+        The version of the metadata defined in the API. This is maintained and enforced by dashboard server. Currently it is 2.0.
 
 
         :param metadata_version: The metadata_version of this ManagementDashboardForImportExportDetails.
@@ -527,7 +534,7 @@ class ManagementDashboardForImportExportDetails(object):
     def ui_config(self):
         """
         **[Required]** Gets the ui_config of this ManagementDashboardForImportExportDetails.
-        JSON that contains user interface options.
+        It defines the visualization type of the widget saved search, the UI options of that visualization type, the binding of data to the visualization.
 
 
         :return: The ui_config of this ManagementDashboardForImportExportDetails.
@@ -539,7 +546,7 @@ class ManagementDashboardForImportExportDetails(object):
     def ui_config(self, ui_config):
         """
         Sets the ui_config of this ManagementDashboardForImportExportDetails.
-        JSON that contains user interface options.
+        It defines the visualization type of the widget saved search, the UI options of that visualization type, the binding of data to the visualization.
 
 
         :param ui_config: The ui_config of this ManagementDashboardForImportExportDetails.
@@ -551,7 +558,7 @@ class ManagementDashboardForImportExportDetails(object):
     def data_config(self):
         """
         **[Required]** Gets the data_config of this ManagementDashboardForImportExportDetails.
-        Array of JSON that contain data source options.
+        It defines how data is fetched. A functional saved search needs a valid dataConfig. See examples on how it can be constructed for various data sources.
 
 
         :return: The data_config of this ManagementDashboardForImportExportDetails.
@@ -563,7 +570,7 @@ class ManagementDashboardForImportExportDetails(object):
     def data_config(self, data_config):
         """
         Sets the data_config of this ManagementDashboardForImportExportDetails.
-        Array of JSON that contain data source options.
+        It defines how data is fetched. A functional saved search needs a valid dataConfig. See examples on how it can be constructed for various data sources.
 
 
         :param data_config: The data_config of this ManagementDashboardForImportExportDetails.
@@ -666,6 +673,30 @@ class ManagementDashboardForImportExportDetails(object):
         :type: list[object]
         """
         self._parameters_config = parameters_config
+
+    @property
+    def features_config(self):
+        """
+        Gets the features_config of this ManagementDashboardForImportExportDetails.
+        Contains configuration for enabling features.
+
+
+        :return: The features_config of this ManagementDashboardForImportExportDetails.
+        :rtype: object
+        """
+        return self._features_config
+
+    @features_config.setter
+    def features_config(self, features_config):
+        """
+        Sets the features_config of this ManagementDashboardForImportExportDetails.
+        Contains configuration for enabling features.
+
+
+        :param features_config: The features_config of this ManagementDashboardForImportExportDetails.
+        :type: object
+        """
+        self._features_config = features_config
 
     @property
     def drilldown_config(self):